trying to update multer config file
